fix: refactor projector
